requirements.txt
requests
urllib3
tqdm
trafilatura
beautifulsoup4
lxml
PyMuPDF
pdfminer.six
